feat/ implementing privilages for admins and users using laravel_gates

+++ app/Http/Controllers/Dashboard/CategoriesController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\CategoryRequest;
 use App\Models\Category;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
 
@@ -17,6 +18,9 @@ class CategoriesController extends Controller
      */
     public function index()
     {
+        if (!Gate::allows('categories.view')) {
+            abort(403);
+        }
         return view('dashboard.categories.index', [
             // SELECT categories.* , parents.name as parent_name FROM categories as a LEFT JOIN categories as parents ON parents.id = categories.parent_id
             'categories' => Category::leftJoin('categories as parents', 'parents.id', '=', 'categories.parent_id')
@@ -37,6 +41,11 @@ class CategoriesController extends Controller
      */
     public function create()
     {
+        // if (!Gate::allows('categories.create')) {
+        //     abort(403);
+        // }
+
+        Gate::authorize('categories.create');
         return view('dashboard.categories.create', ['parents' => Category::all(), 'category' => new Category()]);
     }
 
@@ -61,6 +70,9 @@ class CategoriesController extends Controller
      */
     public function show(Category $category)
     {
+        if (!Gate::allows('categories.view')) {
+            abort(403);
+        }
         return view('dashboard.categories.show', ['category' => $category]);
     }
 
@@ -69,6 +81,8 @@ class CategoriesController extends Controller
      */
     public function edit(string $id)
     {
+        Gate::authorize('categories.update');
+
         $category = Category::find($id);
         if (!$category)
             return redirect(route('dashboard.categories.index'))->with('danger', 'no category found');
@@ -117,6 +131,9 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        if (!Gate::allows('categories.delete')) {
+            abort(403);
+        }
         $category->delete();
         // Category::destroy($id);
         return redirect(route('dashboard.categories.index'))->with('warning', 'category has been deleted');
@@ -124,11 +141,17 @@ class CategoriesController extends Controller
 
     public function trash()
     {
+        if (!Gate::allows('categories.delete')) {
+            abort(403);
+        }
         return view('dashboard.categories.trash', ['categories' => Category::onlyTrashed()->filterBy_name_status(request()->query())->paginate(3)]);
     }
 
     public function restore(string $id)
     {
+        if (!Gate::allows('categories.view')) {
+            abort(403);
+        }
         $category = Category::onlyTrashed()->findOrFail($id);
 
         if (!$category) {
@@ -141,6 +164,9 @@ class CategoriesController extends Controller
 
     public function forceDelete(string $id)
     {
+        if (!Gate::allows('categories.delete')) {
+            abort(403);
+        }
         $category = Category::onlyTrashed()->findOrFail($id);
         if (!$category) {
             return to_route('dashboard.categories.trash')->with('danger', 'category not found');
